+++ src/main/java/nextstep/subway/applicaion/dto/subwayline/SubwayLineResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.applicaion.dto.subwayline;
 
 import lombok.Getter;
 import nextstep.subway.domain.Station;
