+++ src/main/java/nextstep/subway/repository/SubwayLineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.repository;
 
+import nextstep.subway.domain.SubwayLine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SubwayLineRepository extends JpaRepository<SubwayLine, Long> {
