+++ src/test/java/nextstep/subway/acceptance/SubwayLineTest.java
@@ -5,7 +5,7 @@ import io.restassured.path.json.JsonPath;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.applicaion.StationService;
-import nextstep.subway.applicaion.dto.StationRequest;
+import nextstep.subway.applicaion.dto.station.StationRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
